check-version-pinning/requirements.txt
pyyaml == 6.0.2
